# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Khabar News App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /news/feed:
    get:
      operationId: getNewsFeed
      tags: [news]
      summary: Get news feed
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: News feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsFeedResponse"
  /news/trending:
    get:
      operationId: getTrendingNews
      tags: [news]
      summary: Get trending news
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Trending articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsFeedResponse"
  /news/search:
    get:
      operationId: searchNews
      tags: [news]
      summary: Search news
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsFeedResponse"
  /news/{id}:
    get:
      operationId: getArticle
      tags: [news]
      summary: Get article by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleDetail"
  /news/{id}/like:
    post:
      operationId: likeArticle
      tags: [social]
      summary: Like/unlike an article
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
  /news/{id}/save:
    post:
      operationId: saveArticle
      tags: [social]
      summary: Save/unsave an article
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Save toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveResponse"
  /news/{id}/comments:
    get:
      operationId: getComments
      tags: [social]
      summary: Get comments for an article
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentsResponse"
    post:
      operationId: addComment
      tags: [social]
      summary: Add a comment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCommentRequest"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /news/saved:
    get:
      operationId: getSavedArticles
      tags: [news]
      summary: Get saved articles
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Saved articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsFeedResponse"
  /users/{id}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{id}/follow:
    post:
      operationId: followUser
      tags: [users]
      summary: Follow/unfollow a user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Follow toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResponse"
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
  /users/me/profile:
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update own profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        followersCount:
          type: integer
        followingCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - username
        - email
        - displayName
        - followersCount
        - followingCount
        - createdAt
    UserProfile:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            isFollowing:
              type: boolean
          required:
            - isFollowing
    UsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
      required:
        - users
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
    Article:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: string
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        sourceUrl:
          type: string
        sourceName:
          type: string
        category:
          type: string
        publishedAt:
          type: string
        likesCount:
          type: integer
        commentsCount:
          type: integer
        savesCount:
          type: integer
        isLiked:
          type: boolean
        isSaved:
          type: boolean
      required:
        - id
        - title
        - summary
        - content
        - sourceUrl
        - sourceName
        - category
        - publishedAt
        - likesCount
        - commentsCount
        - savesCount
        - isLiked
        - isSaved
    ArticleDetail:
      allOf:
        - $ref: "#/components/schemas/Article"
    NewsFeedResponse:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/Article"
        total:
          type: integer
        page:
          type: integer
        hasMore:
          type: boolean
      required:
        - articles
        - total
        - page
        - hasMore
    LikeResponse:
      type: object
      properties:
        liked:
          type: boolean
        likesCount:
          type: integer
      required:
        - liked
        - likesCount
    SaveResponse:
      type: object
      properties:
        saved:
          type: boolean
        savesCount:
          type: integer
      required:
        - saved
        - savesCount
    Comment:
      type: object
      properties:
        id:
          type: string
        articleId:
          type: string
        userId:
          type: string
        content:
          type: string
        createdAt:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - id
        - articleId
        - userId
        - content
        - createdAt
        - user
    CommentsResponse:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
      required:
        - comments
    AddCommentRequest:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...
    FollowResponse:
      type: object
      properties:
        following:
          type: boolean
      required:
        - following
...