# lib/api-zod/src/generated/types/saveResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khabar News App API
 * OpenAPI spec version: 0.1.0
 */

export interface SaveResponse {
  saved: boolean;
  savesCount: number;
}